requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-filter==22.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
Faker==14.1.0
gunicorn==20.0.4
Markdown==3.4.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
validate-docbr==1.10.0
whitenoise==6.2.0
wrapt==1.12.1
